requirements.txt
fastapi
uvicorn
python-multipart

streamlit
requests

faiss-cpu
sentence-transformers
numpy

ollama

pypdf

python-dotenv

tqdm